+++ NEWS.md
@@ -1,5 +1,23 @@
 # starschemar 1.1.0.9000
 * Errata in data type in help.
+ft_london_boroughs
+ft_usa_states
+ft_usa_city_county_mrs
+ms_where
+usa_cities
+usa_counties
+usa_states
+usa_divisions
+usa_regions
+usa_nation
+uk_london_boroughs
+
+new_geomultistar
+geomultistar
+define_geoattribute
+define_geoattribute_from_attribute
+define_geoattribute_from_layer
+
 
 # starschemar 1.1.0
 * New dimension enrichment features: [enrich_dimension_export()] and [enrich_dimension_import()].