+++ NEWS.md
@@ -1,6 +1,6 @@
-# starschemar 1.1.0.9000
+# starschemar 1.2.0
 * New data sets to enrich dimensions: [ft_london_boroughs], [ft_usa_states] and [ft_usa_city_county].
-* The data set [ms_mrs] has been defined according to the result obtained in the vignette.
+* The data set [ms_mrs] has been defined according to the result obtained in the vignette, so it can be used by other packages.
 * New function to get missing instances to enrich a dimension: [enrich_dimension_import_test()].
 * New functions to get star schema names and a star schema in a constellation: [get_star_schema_names] and [get_star_schema].
 * Improve the structure and content of the vignette.